Revert "Moved data files and login checking to node server to use api"
