--- poule-app/client/src/components/FullButton/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-import './style.css';
-
-const FullButton = (props) => {
-
-    return(
-        <button className="full-btn" onClick={props.onclick}> {props.text} </button>
-    );
-}
-
-export default FullButton;
