--- poule-app/client/src/components/NewPoule/style.css
@@ -1,46 +1,0 @@
-.pop-up-wrapper{
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background: var(--third-bg-color);
-    padding: 5px;
-    border-radius: 10px;
-    width: 90%;
-    margin: auto;
-    box-shadow: 2px 2px 5px 2px var(--second-bg-color);
-}
-
-.pop-up-wrapper h4{
-    margin: 5px 0px;
-}
-
-.pop-up-wrapper .container .content-container form{
-    display: flex;
-    flex-direction: column;
-}
-
-.pop-up-wrapper .container .content-container form label{
-    color: white;
-    font-size: 16px;
-}
-
-.pop-up-wrapper .container .content-container form input[type='text']{
-    background: var(--second-bg-color);
-    padding: 10px;
-    border-radius: 5px;
-    color: white;
-    font-size: 16px;
-    border: none;
-    margin-top: 10px;
-}
-
-.pop-up-wrapper .container .content-container form input[type='submit']{
-    background: var(--acccent-color-a);
-    padding: 10px;
-    color: white;
-    font-size: 16px;
-    border: none;
-    margin-top: 10px;
-    border-radius: 5px;
-}
