--- poule-app/client/src/data/users.json
@@ -1,42 +1,0 @@
-{
-    "users" : [
-        {
-            "id" : 0,
-            "userName" : "Robin",
-            "poules" : [
-                {
-                    "id" : 0
-                },
-                {
-                    "id": 2
-                }
-                
-            ],
-            "settings" : [
-
-            ]
-        },
-        {
-            "id" : 1,
-            "userName" : "Niek",
-            "poules" : [
-                {
-                    "id" : 1
-                }
-            ],
-            "settings" : [
-
-            ]
-        },
-        {
-            "id" : 2,
-            "userName" : "Jeroen",
-            "poules" : [
-
-            ],
-            "settings" : [
-
-            ]
-        }
-    ]
-}
