+++ poule-app/src/pages/home/index.js
@@ -1,0 +1,11 @@
+import react from "react";
+
+function Home() {
+    return (
+        <div className="page-content page-home">
+            <h2>This is the homepage</h2>
+        </div>
+    );
+}
+
+export default Home;
