+++ poule-app/src/pages/poules/index.js
@@ -1,0 +1,11 @@
+import react from "react";
+
+function Poules(){
+    return(
+        <div className="page-content page-poules">
+            <h2>This is the poules page</h2>
+        </div>
+    );
+}
+
+export default Poules;
