+++ poule-app/src/pages/profile/index.js
@@ -1,0 +1,11 @@
+import react from "react";
+
+const Profile = () => {
+    return (
+        <div className="page-content page-home">
+            <h2>This is the profile page</h2>
+        </div>
+    );
+}
+
+export default Profile;
